scripts/requirements.txt
supabase==2.10.0
python-dotenv==1.0.0
websockets>=13.0
